feat: render text from file and scroll (READ long msg)
Like the previous commit, this is a 'first draft':
the scroll can still 'overflow' and the rendering
could be done better. This just me getting something
that works first.

For some reason ansi's escape code for screen clearing
gave me a lot of trouble by scrolling instead of clearing
properly; for this reason I had to use the same technique
Tsoding uses in Noed (fill up a buffer of terminal-window-size
with space-ch and then write the text in it). I plan to
change this as I already took too much from his code.