feat: render text, edit and save file (READ long msg)
At the moment if you open a file and it's empty
you can type and save with <Esc> and exit
the editor with <q>.
While writing the editor is basically in 'unbuffered mode',
you cannot:
	- navigate the file, you can only go forward by typing
	  new chars
	- delete text
	- scroll text

Also at the moment you cannot write on a non empty file
but you can render its first n-terminal_rows lines.
The reason being that right now each Piece is added
to the piece-table **AFTER** the 'original' Piece pointing
to the file-buf's full content.

This is because, once again, I'm focusing on getting
something that works first, meaning the ability to:
	- render file contents
	- edit the file
	- save the file
To do the second thing in the list I needed to add pieces
from the add-buf after the original Piece.
So it still lacks a lot of fundamental features but right
now Fred is much closer to being a basic editor than before.