fix: text-rendering from multiple pieces (read long msg)
For some reason the pointer comparison I used
in get_text_from_piece_table() to check if
I reached the piece-table's end doesn't work properly.

I was using this method thinking that it would be more
efficient to not have a nested loop in a function
that gets called so frequently, especially
when the table will grow substantially, even
though I had and still have no data to back that up.

Now it uses a nested loop, but it will get
updated if I figure out a more efficient way.