IMPORTANT feat + fix: add cursor scroll; fix line-num's rendering of last line in window (read long msg)
- Cursor scroll:
Scrolling handles wrapped lines as well. The code
will get cleaned up later.

- Line-num rendering:
if the last line in the terminal window
was longer than the amount of rows left in the window (had to
be a really long line), the line-num wouldn't get rendered
and other line-nums would get misplaced.
This was because I was waiting to reach a newline character
to calculate the the line-num. The problem was that
I was also returning from the function if I had reached
the end of the window while rendering.
So if the last line was too long, its line-num wouldn't get
rendered.