feat (kinda): edit file (only append text, READ long msg)
The editing pushes pieces at the end of the
table; if you open a non-empty file the edits
won't happen in the middle of the text.
So basically you can only edit and empty file
(but not save, i disabled it for the moment so
i don't fuck something up).

It's after-cursor editing only. If the new editing
happens at the last edited spot, no new pieces
will be pushed to the table, it will increment the
length of the last-edit's piece.

This was to test the rendering and the cursor
movement across and along the lines.
It already highlighted that the lines scrolling
should be independant from FRED_cursor_move()
as it should also happen when editing (im dumb i know).