fix + refactor: line-nums (again, READ long msg); make macros for text-insertions
line-nums: turns out that on empty files (hence empty table)
line-nums were rendered (again) only after typing the first char
after '\n'. im so done with this shi
I've decided only now to do the dumbest thing possible (rendering
 1st line-num regardless and the others when finding a '\n'),
because i wanted the line-nums code to be written all in the same
place. But now I think, not sure, that this is better and even
faster (?), since less checks are needed, but idk.