IMPORTANT: rm generate_tests.lua (read long msg)
The idea of sending random keys to Neovim through
nvim_feedkeys() is unusable because I cannot
'take a screenshot' of the edited file right
after sending one key, which was the entire
premise of the test.

Also, <Esc> in Neovim moves the cursor and I would have
to use 'a' instead of 'i' to insert, but this is a small
thing.