IMPORTANT fix: (READ msg!) heap corruption in insert/delete by memmoving freed pointers
During insertion/deletion I was passing pointers
to the piece-table, that were already freed by realloc,
to memmove.

This went unnoticed because I was allocating a lot of
memory for the piece-table from the start.
I only noticed after I tried to run fred_test_big/
Now offsets instead of pointers are used.

NOTE (MORE BUGS): insertion/deletion IS STILL BROKEN, as it
does not manage to pass fred_test_big/;
all the other 'smaller' tests are passed successfully.

This is caused by char inserted at the wrong place, only
after a lot of edits.
It has to do with the cursor, which right now
can go off-boundary by +1 when scrolling down.