IMPORTANT refactor: use offsets to snapshots file (read msg)
I have no idea why the fuck was i allocating
memory for each snap instead of reusing to the file.

Cursor pos retrieval from snapshots is for now
unused as it's kinda useless:
any mismatch between cursor coordinates is already reported
by the mismatch in text, so it doesn't really
tell you anything new.
It would be more useful to know after which nav-keypress
the mismatch is found.