IMPORTANT fix: (READ msg) curs-col not 'fixed' on vert-scroll in key-gen; curs in snap is now curs before edit
When scrolling in Fred, if cursor-col < next-line-len,
we set it to be next-line-len.

In this script we were checking and fixing the cursor-col
ONLY with the line at the next-random-position, instead
of doing it for each line IN BETWEEN the current and
the next-random-position.

This generated the failing tests like 'fred_test_big'