feat: introduce faster(?) text retrieval for rendering (read msg)
Before we were parsing the entire the table char-by-char, from
the start all the time, just to reach to the 1st line to be rendered.
This meant parsing a lot of useless text/pieces, especially
when you're deep down in the file.

Now instead we find the piece containg said line by using
the same idea of the text-editing functions:
using the cached lines-lengths to get an 'absolute' offset
of the 1st line to render,
and then parsing the table until the sum of the previous
pieces's lenghts is either equal or greater then said offset.

This way we can parse much less data, since we can (pretty
much) assume there will always be less pieces than text-chars.

Now I profiled this with gprof on a 1000+ lines file and
it was 1 order of magnitude faster (0.10ms VS  0.01ms),
but I haven't profiled this on a table with large piece-count,
hence the (?).

NOTE: this parses the EOL at endoffile in posix compliant files as
an actual line, so the line-num is shown.