feat: add text highlighting (BREAKS rendering read msg)
We now store and highlight the full
piece-table's text in a dynamic array
and use that for rendering instead of
parsing the table at each frame.

At the moment the color-highlight
is fully hardcoded.

Broken rendering:
the text-building/highlighting works just fine,
the issue is with FRED_get_text_to_render() and
the text placement in the TermWin-array:
the moment there's an escape-code the text
after it gets shifted back.