fix: highlighting (render still broken, diff reason, read msg)
Storing the escape codes inside the built-text
didn't work out, so for now in the built-text
we will just flag the each keyword start with
its negative length.
We then store the row,col and length of the keywords
that need to be shown in a dynamic array, using
them to print the highlighted keywords ON TOP
the already drawn text.

This method is not memory expensive and slow,
sometimes we can see the cursor flicker if you
spam a key in insert mode.

BROKEN RENDERING: it's caused by using the NORMAL
lines-length to find the offset first line to render.
This lines-length do not take into account the
keywords byte-flag, so you're starting to render
many characters earlier.